refactor/016: moved the ResponseResult to a common block and added the ability to change the color in the editor theme
