+++ src/features/ResponseResult/ui/ResponseResult.module.scss
@@ -1,8 +1,21 @@
-.result {
-  max-height: 250px;
-  overflow: auto;
-  scrollbar-width: thin;
-  scrollbar-color: var(--main-color-3) var(--main-color-4);
-  padding: 2px;
-  font-size: 14px;
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  max-width: 820px;
+  margin: 0 auto;
+  padding: 10px;
+  border: 1px solid var(--main-color-1);
+  border-radius: 4px;
+  margin-top: 20px;
+}
+
+.status_text {
+  color: var(--main-color-2);
+  color: var(--orange);
+  font-size: 16px;
+  margin-right: auto;
+  margin-bottom: 6px;
 }
