+++ src/widgets/GraphQl/ui/GraphQlPlayground.module.scss
@@ -67,11 +67,3 @@
   color: var(--magenta);
   margin: 20px auto;
 }
-
-.status_text {
-  color: var(--main-color-2);
-  color: var(--orange);
-  font-size: 16px;
-  margin-right: auto;
-  margin-bottom: 6px;
-}
