+++ src/widgets/GraphQl/ui/GraphQlPlayground.tsx
@@ -10,9 +10,9 @@ import {
   DEFAULT_URL_GRAPHQL_EXAMPLE,
 } from 'shared/constants';
 import { encode64 } from 'shared/lib/dataConverters';
-import { PropsArea } from './components';
 import style from './GraphQlPlayground.module.scss';
 import { Path } from 'shared/types/path';
+import { PropsArea } from './PropsArea/PropsArea';
 
 const GraphQlPlayground = ({ children }: { children?: ReactNode }) => {
   const navigate = useRouter();
