+++ src/widgets/GraphQl/ui/PropsArea/PropsArea.module.scss
