--- src/widgets/GraphQl/ui/components/index.ts
@@ -1,2 +1,0 @@
-export { myTheme } from './EditorView';
-export { PropsArea } from './PropsArea';
